requirements.txt
feedparser
pyyaml
scipy
scikit-learn
